added range map

+++ src/main/java/fr/bananasmoothii/rulesgeneration/suggestions/SimpleSuggestion.java
@@ -25,7 +25,7 @@ public class SimpleSuggestion extends Suggestion {
     }
 
     @Override
-    public double shouldFollow() {
+    public float shouldFollow() {
         return what.rarity;
     }
 }
